changed collision system
- deleted all the tree and lamp post objects
- added new sorting layer: tall objects
From now on, add the "base" of a tall object (bottom of a tree, bottom of a lamp post) on the "collision" tile map, which is set to the "collision" sorting layer. Add the rest of the tall object on the "tall object" tilemap, set to the "tall object" sorting layer.

-custom physics shapes for the bases can be created in the tileset Sprite Editor -> custom physics shape